refactor catalog to merge v2 changes into v1 interface
